Use checkered bg for image export previews
